+++ .prettierrc
@@ -3,5 +3,6 @@
   "singleQuote": true,
   "tabWidth": 2,
   "useTabs": false,
-  "trailingComma": "none"
+  "trailingComma": "none",
+  "printWidth": 120
 }
